% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGallery.R
\name{artInputTag}
\alias{artInputTag}
\title{Create an Art Input Tag from the tag string}
\usage{
artInputTag(tag)
}
\arguments{
\item{tag}{The tag string. The tag can have up to three parts separated by colon:
\code{moduleLabel:instanceLabel:processName}. Usually, you only need the \code{moduleLabel}.
Sometimes you may need the \code{instanceLabel}. Rarely will you need the \code{processName}.}
}
\value{
The python representation of the C++ \code{art::InputTag} object
}
\description{
Create an Art Input Tag from the tag string
}
\examples{
\dontrun{
ghostTag <- artInputTag('artg4:GhostCylinderDetector')
calTag <- artInputTag('artg4')
}
}
