% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{createFileVector}
\alias{createFileVector}
\title{Create the file vector}
\usage{
createFileVector(fileList)
}
\arguments{
\item{fileList}{A list of files to process}
}
\value{
The python representation of the std::vector<std::string>
}
\description{
The files must be presented to gallery as a std::vector
}
