% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGallery.R
\name{createReaderClass_from_file}
\alias{createReaderClass_from_file}
\title{Create a reader class from a file}
\usage{
createReaderClass_from_file(f)
}
\arguments{
\item{f}{}
}
\value{
The main python environment. Use it to extract the class. See Vignette
}
\description{
See Vignette for more information
}
