% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGallery.R
\name{createReaderClass_from_string}
\alias{createReaderClass_from_string}
\title{Create a reader class from a string}
\usage{
createReaderClass_from_string(s)
}
\arguments{
\item{s}{}
}
\value{
The main python environment. Use it to extract the class. See Vignette
}
\description{
See Vignette for more information
}
