% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGallery.R
\name{galleryReader_df}
\alias{galleryReader_df}
\title{Extract the data from a reader as a data frame}
\usage{
galleryReader_df(reader)
}
\arguments{
\item{reader}{THe reader from which to extract the data frame}
}
\value{
A data frame
}
\description{
Extract the data from a reader as a data frame
}
