% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGallery.R
\name{getGalleryData}
\alias{getGalleryData}
\title{Read data from Gallery}
\usage{
getGalleryData(theFile, reader)
}
\arguments{
\item{theFile}{A file or list of files to process}

\item{reader}{A reader or list of readers to use for processing}
}
\value{
Timing information
}
\description{
See the vignette for more information
}
