% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readerClassSkel.R
\name{readerClassSkel}
\alias{readerClassSkel}
\title{Create a Gallery Reader Skeleton}
\usage{
readerClassSkel(extractClass, fillClass = NULL, wrapExtractVector = TRUE,
  wrapFillVector = NULL, writeFile = NULL)
}
\arguments{
\item{extractClass}{- The name of the class, with namespace, that is extracted from Gallery}

\item{fillClass}{- The name of the class to be filled, Usually same as extractClass (default NULL does this)}

\item{wrapExtractVector}{- If TRUE, wrap the extract class in a vector}

\item{wrapFillVector}{- if TRUE, wrap the fill class in a vector (default, NULL, is to copy wrapExtractVector)}

\item{writeFile}{- Write the text to a file named \code{writeFile}}
}
\value{
[Invisibly] the text of the class
}
\description{
This function aids you in writing a Gallery Reader class.
}
