% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{readersFromList}
\alias{readersFromList}
\title{Create a reader object that consists of a list of readers}
\usage{
readersFromList(listOfReaderObjects)
}
\arguments{
\item{listOfReaderObjects}{The list of readers to put in the object}
}
\value{
A python representation of the GalleryReaders object
}
\description{
Create a reader object that consists of a list of readers
}
