% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGallery.R
\name{useDataProduct}
\alias{useDataProduct}
\title{Use an Art Data Product}
\usage{
useDataProduct(dataProductString)
}
\arguments{
\item{dataProductString}{The data product (perhaps with std::vector) that you want to use}
}
\value{
Nothing
}
\description{
Before you read data, you must declare what data products you want to read. You pass in the
full C++ name of the object, including wrapped in \code{std::vector<>} if necessary.
Note that you may need to look at your console output for error messages.
}
\examples{
\dontrun{
useDataProduct('std::vector<gm2truth::GhostDetectorArtRecord>')
useDataProduct('gm2info::SpecialEventInfoRecord')
}
}
