% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{xrootify}
\alias{xrootify}
\title{Add the Fermilab public dCache URI to the front of file paths}
\usage{
xrootify(files)
}
\arguments{
\item{files}{A list of file locations (should start with /pnfs)}
}
\value{
The file list with the Fermilab public dCache URI prepended
}
\description{
Add the Fermilab public dCache URI to the front of file paths
}
\examples{
xrootify("/pnfs/GM2/scratch/users/lyon/myFile.root")
}
